Add top new scouts export to Admin
